gui: refactor GUI sta interface and Qt modules to allow for better sharing for CTS clock trees
