+++ src/gui/CMakeLists.txt
@@ -42,6 +42,7 @@ if (Qt5_FOUND AND BUILD_GUI)
     src/highlightGroupDialog.cpp
     src/selectHighlightWindow.cpp
     src/staGui.cpp
+    src/staGuiInterface.cpp
     src/timingWidget.cpp
     src/drcWidget.cpp
     src/ruler.cpp
